Moved test into __test__ folder.
