+++ src/__tests__/index.test.js
@@ -1,5 +1,5 @@
-import * as JobListings from './index';
-import { getBuilditJobPostings } from './buildit';
+import * as JobListings from '../index';
+import { getBuilditJobPostings } from '../buildit';
 
 describe('fetch-job-listings public API', () => {
   test('contains getBuilditJobPostings() function', () => {
